requirements.txt
Flask==2.0.1
requests==2.25.1
python-dotenv==0.19.0
Flask-Session==0.4.0